+++ R/data_to_wide.R
@@ -25,9 +25,6 @@
 #' missing values in the new columns created.
 #' @param verbose Toggle warnings.
 #' @param ... Not used for now.
-#' @param colnames_from Deprecated. Use `names_from` instead.
-#' @param rows_from Deprecated. Use `id_cols` instead.
-#' @param sep Deprecated. Use `names_sep` instead.
 #'
 #' @return If a tibble was provided as input, `reshape_wider()` also returns a
 #' tibble. Otherwise, it returns a data frame.
@@ -93,28 +90,7 @@ data_to_wide <- function(data,
                          names_glue = NULL,
                          values_fill = NULL,
                          verbose = TRUE,
-                         ...,
-                         colnames_from,
-                         rows_from,
-                         sep) {
-  if (!missing(colnames_from)) {
-    .is_deprecated("colnames_from", "names_from")
-    if (is.null(names_from)) {
-      names_from <- colnames_from
-    }
-  }
-  if (!missing(rows_from)) {
-    .is_deprecated("rows_from", "id_cols")
-    if (is.null(id_cols)) {
-      id_cols <- rows_from
-    }
-  }
-  if (!missing(sep)) {
-    .is_deprecated("sep", "names_sep")
-    if (is.null(names_sep)) {
-      names_sep <- sep
-    }
-  }
+                         ...) {
   if (is.null(id_cols)) {
     id_cols <- setdiff(names(data), c(names_from, values_from))
   }
@@ -218,28 +194,26 @@ data_to_wide <- function(data,
   if (!is.null(values_fill)) {
     if (length(values_fill) == 1L) {
       if (is.numeric(new_data[[values_from]])) {
-        if (!is.numeric(values_fill)) {
-          insight::format_error(paste0("`values_fill` must be of type numeric."))
-        } else {
+        if (is.numeric(values_fill)) {
           new_data <- convert_na_to(new_data, replace_num = values_fill)
+        } else {
+          insight::format_error(paste0("`values_fill` must be of type numeric."))
         }
       } else if (is.character(new_data[[values_from]])) {
-        if (!is.character(values_fill)) {
-          insight::format_error(paste0("`values_fill` must be of type character."))
-        } else {
+        if (is.character(values_fill)) {
           new_data <- convert_na_to(new_data, replace_char = values_fill)
+        } else {
+          insight::format_error(paste0("`values_fill` must be of type character."))
         }
       } else if (is.factor(new_data[[values_from]])) {
-        if (!is.factor(values_fill)) {
-          insight::format_error(paste0("`values_fill` must be of type factor."))
-        } else {
+        if (is.factor(values_fill)) {
           new_data <- convert_na_to(new_data, replace_fac = values_fill)
+        } else {
+          insight::format_error(paste0("`values_fill` must be of type factor."))
         }
       }
-    } else {
-      if (verbose) {
-        insight::format_error("`values_fill` must be of length 1.")
-      }
+    } else if (verbose) {
+      insight::format_error("`values_fill` must be of length 1.")
     }
   }