+++ R/data_xtabulate.R
@@ -312,8 +312,39 @@ print_html.dw_data_xtabulates <- function(x, big_mark = NULL, ...) {
 }
 
 
-.validate_table_weights <- function(weights, x) {
-  if (!is.null(weights)) {
+.validate_table_weights <- function(weights, x, weights_expression = NULL) {
+  # exception: for vectors, if weighting variable not found, "weights" is NULL.
+  # to check this, we further need to check whether a weights expression was
+  # provided, e.g. "weights = iris$not_found" - all this is only relevant when
+  # weights is NULL
+  if (is.null(weights)) {
+    # possibly misspelled weights-variables for default-method ----------------
+    # -------------------------------------------------------------------------
+
+    # do we have any value for weights_expression?
+    if (!is.null(weights_expression) &&
+      # due to deparse() and substitute, NULL becomes "NULL" - we need to check for this
+      !identical(weights_expression, "NULL") &&
+      # we should only run into this problem, when a variable from a data frame
+      # is used in the data_tabulate() method for vectors - thus, we need to check
+      # whether the weights_expression contains a "$" - `iris$not_found` is "NULL"
+      # we need this check, because the default-method of data_tabulate() is called
+      # from the data.frame method, where `weights = weights`, and then,
+      # deparse(substitute(weights)) is "weights" (not "NULL" or "iris$not_found"),
+      # leading to an error when actually all is OK (if "weights" is NULL)
+      # Example:
+      #> efc$weights <- abs(rnorm(n = nrow(efc), mean = 1, sd = 0.5))
+      # Here, efc$wweight is NULL
+      #> data_tabulate(efc$c172code, weights = efc$wweight)
+      # Here, wweight errors anyway, because object "wweight" is not found
+      #> data_tabulate(efc$c172code, weights = wweight)
+      grepl("$", weights_expression, fixed = TRUE)) {
+      insight::format_error("The variable specified in `weights` was not found. Possibly misspelled?")
+    }
+  } else {
+    # possibly misspecified weights-variables for data.frame-method -----------
+    # -------------------------------------------------------------------------
+
     if (is.character(weights)) {
       # If "weights" is a character string, must be of length 1
       if (length(weights) > 1) {